# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ClashLayoutsHub API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface Base {
  id: string;
  title: string;
  slug: string;
  townhall: number;
  base_type: string;
  image_url: string;
  layout_link: string;
  description: string;
  views?: number;
  copies?: number;
  report_count?: number;
  is_active?: boolean;
  key_features?: string[];
  best_against?: string[];
  win_rate?: number;
  difficulty?: string;
  rating_avg?: number;
  rating_count?: number;
  health_score?: number;
  approved?: boolean;
  created_at?: string;
  updated_at?: string;
}

export interface CreateBaseBody {
  title: string;
  slug?: string;
  townhall: number;
  base_type: string;
  image_url: string;
  layout_link: string;
  description: string;
  key_features?: string[];
  best_against?: string[];
  win_rate?: number;
  difficulty?: string;
}

export interface UpdateBaseBody {
  title?: string;
  townhall?: number;
  base_type?: string;
  image_url?: string;
  layout_link?: string;
  description?: string;
  key_features?: string[];
  best_against?: string[];
  win_rate?: number;
  difficulty?: string;
  is_active?: boolean;
  approved?: boolean;
}

export interface BasesListResponse {
  bases: Base[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface SiteStats {
  totalBases: number;
  totalViews: number;
  totalCopies: number;
  totalBlogPosts: number;
}

export interface TownHallSummary {
  townhall: number;
  count: number;
  topType: string;
}

export interface CopyResponse {
  copies: number;
  message: string;
}

export interface TodayCopiesResponse {
  todayCopies: number;
}

export interface BlogPost {
  id: string;
  title: string;
  slug: string;
  content: string;
  excerpt: string;
  featured_image: string;
  views?: number;
  meta_title?: string;
  meta_description?: string;
  published?: boolean;
  published_at?: string;
  created_at?: string;
}

export interface CreateBlogPostBody {
  title: string;
  slug?: string;
  content: string;
  excerpt: string;
  featured_image: string;
  meta_title?: string;
  meta_description?: string;
  published?: boolean;
}

export interface BlogListResponse {
  posts: BlogPost[];
  total: number;
  page: number;
  totalPages: number;
}

export interface Comment {
  id: string;
  base_id: string;
  user_name: string;
  content: string;
  rating?: number;
  approved?: boolean;
  created_at?: string;
}

export interface CreateCommentBody {
  user_name: string;
  content: string;
  rating?: number;
}

export interface Report {
  id: string;
  base_id: string;
  base_title?: string;
  reason: string;
  reporter_ip?: string;
  created_at?: string;
}

export interface CreateReportBody {
  base_id: string;
  reason: string;
}

export interface ReportsListResponse {
  reports: Report[];
  total: number;
  page: number;
  totalPages: number;
}

export interface Submission {
  id: string;
  townhall?: number;
  base_type?: string;
  layout_link: string;
  description?: string;
  image_url?: string;
  status: string;
  admin_notes?: string;
  submitted_at?: string;
}

export interface CreateSubmissionBody {
  townhall?: number;
  base_type?: string;
  layout_link: string;
  description?: string;
  image_url?: string;
}

export type UpdateSubmissionBodyStatus =
  (typeof UpdateSubmissionBodyStatus)[keyof typeof UpdateSubmissionBodyStatus];

export const UpdateSubmissionBodyStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface UpdateSubmissionBody {
  status: UpdateSubmissionBodyStatus;
  admin_notes?: string;
}

export interface SubmissionsListResponse {
  submissions: Submission[];
  total: number;
  page: number;
  totalPages: number;
}

export type DashboardAnalyticsBasesByTownhallItem = {
  townhall: number;
  count: number;
};

export interface DashboardAnalytics {
  totalBases: number;
  totalViews: number;
  pendingSubmissions: number;
  activeReports: number;
  basesByTownhall: DashboardAnalyticsBasesByTownhallItem[];
}

export interface ViewsAnalyticsEntry {
  date: string;
  views: number;
}

export type ListBasesParams = {
  /**
   * @minimum 3
   * @maximum 18
   */
  townhall?: number;
  base_type?: string;
  difficulty?: string;
  sort?: ListBasesSort;
  page?: number;
  limit?: number;
  search?: string;
};

export type ListBasesSort = (typeof ListBasesSort)[keyof typeof ListBasesSort];

export const ListBasesSort = {
  latest: "latest",
  most_viewed: "most_viewed",
  top_rated: "top_rated",
  most_copied: "most_copied",
} as const;

export type ListBlogPostsParams = {
  page?: number;
  limit?: number;
};

export type ListReportsParams = {
  page?: number;
};

export type ListSubmissionsParams = {
  status?: ListSubmissionsStatus;
  page?: number;
};

export type ListSubmissionsStatus =
  (typeof ListSubmissionsStatus)[keyof typeof ListSubmissionsStatus];

export const ListSubmissionsStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;