# lib/api-zod/src/generated/types/createBaseBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ClashLayoutsHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBaseBody {
  title: string;
  slug?: string;
  townhall: number;
  base_type: string;
  image_url: string;
  layout_link: string;
  description: string;
  key_features?: string[];
  best_against?: string[];
  win_rate?: number;
  difficulty?: string;
}