# lib/api-zod/src/generated/types/createBlogPostBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ClashLayoutsHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBlogPostBody {
  title: string;
  slug?: string;
  content: string;
  excerpt: string;
  featured_image: string;
  meta_title?: string;
  meta_description?: string;
  published?: boolean;
}