# lib/api-zod/src/generated/types/todayCopiesResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ClashLayoutsHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TodayCopiesResponse {
  todayCopies: number;
}